Refactor observability infrastructure to simplify DI into involved components.
We now ship a ObservationContext bean that captures both a ObservationRegistry, (nullable) MeterRegistry and Environment to avoid having to pass around all of them to the various components that end up needing access to them.

Extracted the observation handling part in ModuleEntryInterceptor into a nested class to prepare for composition later on. Reduce visibility of as many components as possible to later decide about which part to expose explicitly.