GH-1250 - Register runtime hints for JSONPath internals for native image builds.
Our AOT support generates a JSON file containing application module metadata for runtime usage. That file is evaluated using JSONPath expressions and that evaluation triggers reflection inside the JSONPath library. We now register reflection hints allow the invocation of the constructors of the necessary types.

This can be removed as soon as https://github.com/json-path/JsonPath/issues/1042 is fixed.