GH-1216 - Avoid referring to the source class in ModuleTypeExcludeFilter.equals(…) and ….hashCode().
MTEF's equals(…) and hashCode() methods is depended on by the application context caching facilities of Spring Framework's test context framework. Previously, it referred to the class under test in those methods, thus effectively disabling the cache, even if the application module test bootstrap setup was identical but used on different test classes.

We now avoid this by solely referring to the equals(…) and hashCode() methods of ModuleTestExecution, which in turn only considers the base package and the @ApplicationModuleTest annotation.