GH-1280 - Prevent named interfaces from leaking into parent modules.
We need to forward the exclusions we collect to bootstrap nested modules into the named interface detection as otherwise the ones declared for nested modules are also detected for parent ones. This is now achieved by allowing JavaPackage instances to be created with exclusions and the module base package be prepared with those before the named general module setup kicks off.