GH-1287 - Support for runtime application module verification.
The spring-modulith-runtime artifact now exposes a `spring.modulith.runtime.verification-enabled` property to initiate application module arrangement verification on application startup. The ApplicationModules instance is bootstrapped asynchronously and any failure is elevated onto the main application thread in the post-singleton-initialization phase. We also register a FailureAnalyzer implementation to give a bit of additional guidance on how to deal with the problem.